extract image processing logic
